requirements.txt
streamlit>=1.0
torch>=1.10
torchvision>=0.11
Pillow>=8.0
numpy>=1.19
